refactor: pull in vsock.go module from firecracker-containerd
